+++ packages/scene-astro/src/index.ts
@@ -1,3 +1,2 @@
-export {interactiveSceneComponentManifest} from './editor'
 export {serializeSceneConfig} from './serialization'
 export type {InteractiveSceneDefinition, SceneInteractionMode, SceneQuality} from '@ooopsstudio/scene-core'